@@ -8,5 +8,8 @@ dependencies:
   pointycastle: ^3.6.2
   typed_data: ^1.3.1
 
+dev_dependencies:
+  test: ^1.5.1
+
 environment:
-  sdk: ">=1.19.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
